+++ qiskit/providers/ibmq/api/rest/root.py
@@ -78,6 +78,7 @@ class Api(RestAdapterBase):
             self,
             limit: int = 10,
             skip: int = 0,
+            descending: bool = True,
             extra_filter: Dict[str, Any] = None
     ) -> List[Dict[str, Any]]:
         """Return a list of jobs statuses.
@@ -85,6 +86,7 @@ class Api(RestAdapterBase):
         Args:
             limit: maximum number of items to return.
             skip: offset for the items to return.
+            descending: whether the jobs should be in descending order.
             extra_filter: additional filtering passed to the query.
 
         Returns:
@@ -92,8 +94,10 @@ class Api(RestAdapterBase):
         """
         url = self.get_url('jobs_status')
 
+        order = 'DESC' if descending else 'ASC'
+
         query = {
-            'order': 'creationDate DESC',
+            'order': 'creationDate ' + order,
             'limit': limit,
             'skip': skip,
         }