+++ qiskit/providers/ibmq/ibmqbackendservice.py
@@ -104,6 +104,7 @@ class IBMQBackendService(SimpleNamespace):
             end_datetime: Optional[datetime] = None,
             job_tags: Optional[List[str]] = None,
             job_tags_operator: Optional[str] = "OR",
+            descending: bool = True,
             db_filter: Optional[Dict[str, Any]] = None
     ) -> List[IBMQJob]:
         """Return a list of jobs from the API.
@@ -141,6 +142,8 @@ class IBMQBackendService(SimpleNamespace):
                 * If "OR" is specified, then a job only needs to have any
                     of the tags specified in ``job_tags`` to be included.
                 Default: OR.
+            descending: if True, return the jobs in descending order of the job
+                creation date (newest first). If False, return in ascending order. Default: True.
             db_filter: `loopback-based filter
                 <https://loopback.io/doc/en/lb2/Querying-data.html>`_.
                 This is an interface to a database ``where`` filter. Default: None.
@@ -232,7 +235,8 @@ class IBMQBackendService(SimpleNamespace):
 
         while True:
             job_page = self._provider._api.list_jobs_statuses(
-                limit=current_page_limit, skip=skip, extra_filter=api_filter)
+                limit=current_page_limit, skip=skip, descending=descending,
+                extra_filter=api_filter)
             job_responses += job_page
             skip = skip + len(job_page)